Move getpid to standalone function
