+++ kernel/src/user_program/syscall.rs
@@ -1,7 +1,7 @@
 // https://docs.google.com/document/d/1qMMU73HW541wME00Ngl79ou-kQ23zzTlGXJYo9FNh5M
 
 use crate::mem::user::check_and_copy_user_memory;
-use crate::system::{unwrap_system, unwrap_system_mut};
+use crate::system::{unwrap_system_mut, running_thread_pid, running_thread_ppid};
 use crate::threading::scheduling::{scheduler_yield_and_continue, scheduler_yield_and_die};
 use crate::threading::thread_control_block::ThreadControlBlock;
 use crate::threading::thread_functions;
@@ -70,31 +70,13 @@ pub extern "C" fn handler(syscall_number: usize, arg0: usize, arg1: usize, arg2:
             scheduler_yield_and_die();
         }
         SYS_GETPID => {
-            let tcb = unsafe {
-                unwrap_system()
-                    .threads
-                    .running_thread
-                    .as_ref()
-                    .expect("Why is nothing running?")
-                    .as_ref()
-            };
-            tcb.pid as isize
+            running_thread_pid() as isize
         }
         SYS_NANOSLEEP => {
             todo!("nanosleep syscall")
         }
         SYS_GETPPID => {
-            let tcb = unsafe {
-                unwrap_system()
-                    .threads
-                    .running_thread
-                    .as_ref()
-                    .expect("Why is nothing running?")
-                    .as_ref()
-            };
-            let process_table = unsafe { &unwrap_system().process.table };
-            let pcb = process_table.get(tcb.pid).unwrap();
-            pcb.ppid as isize
+            running_thread_ppid() as isize
         }
         SYS_SCHED_YIELD => {
             scheduler_yield_and_continue();
